2020.12.02: * fixed: Utilities: x64 configuration build fixup * new: Utilities: x64 configuration * new: Utilities: Born's WSH dynacall 32-bit dll (http://www.borncity.com/web/WSHBazaar1/WSHDynaCall.htm ) port to msvc2015 to enable Win32 calls from vbs scripts
git-svn-id: svn+ssh://svn.code.sf.net/p/contools/contools/trunk@485 699b1267-f50e-4533-8e1b-323e5631ef88

+++ Utilities/src/_common/printf.hpp
@@ -58,7 +58,7 @@ namespace {
 						strcat(var_buffer, "}");
 
 						// {N}
-						const int var_vn_len = strlen(var_buffer);
+						const size_t var_vn_len = strlen(var_buffer);
 						const int var_vn = strncmp(p, var_buffer, var_vn_len);
 						if (!var_vn) {
 							parsed_str.append(last_offset, p);
@@ -73,7 +73,7 @@ namespace {
 							strcat(var_buffer, int_to_str_buffer);
 							strcat(var_buffer, "hs}");
 
-							const int var_vn_len = strlen(var_buffer);
+							const size_t var_vn_len = strlen(var_buffer);
 							const int var_vn = strncmp(p, var_buffer, var_vn_len);
 							if (!var_vn) {
 								parsed_str.append(last_offset, p);
